Code parsing infrastructure, generation of TOCs for Sphinx (#145)
This includes a parser for Matlab source code files (not general purpose, but with the conventions used in RepLAB).

The `help` command has be revamped using that parser and provides user friendly navigation of RepLAB classes, functions and methods.

RepLAB now generates the DocTests using this parser, with support of additional features (multiline commands etc).

We also process the source code to add TOCs in documentation comments as to enrich the Sphinx API output.

Co-authored-by: jdbancal <jdbancal@gmail.com>
Co-authored-by: denisrosset <physics@denisrosset.com>