Tutorial on the eigenvalue problem for a vibrating square membrane
This tutorial illustrates further how RepLAB can be used to perform numerical studies by employing group theoretical methods. Our toy model is a vibrating square membrane, the region of which we discretize by a 9-point lattice and we also make use of the symmetries of the Laplacian operator to considerably simplify the problem, leading to block diagonalization of the associated matrix.