Algorithm to generate symmetry-adapted basis for symmetrization
This is a function that uses the algorithm from Faessler and Stiefel (also Serre's projection formulas) to generate symmetry-adapted basis to decompose a given representation. A choice has been made to label the vectors in a way that a matrix with the symmetry of the representation is brought into block-diagonal form.