New framework for non-permutation groups
This cleans up the hackish construction for non-permutation finite groups. It introduces the concept of a type, which is a possibly non-finite group from which finite subgroups are constructed.

Examples of finite group types include:

- Permutation over `n` elements. In RepLAB, permutation groups over different domain sizes have a different type.
- Signed permutation over `-n,..,-1, 1,..,n`. One type per `n`.
- `n x n` matrices with coefficients in the cyclotomic field. One type per `n`.

Types can be composed: for example, a direct product of types is again a type.

Some methods in finite groups, like `.closure`, accept not only group elements as arguments, but also type elements. This is a way to enlarge the group in a principled way (and "type-safe" way, even though Matlab is not statically typed)

Co-authored-by: Joscelyn <joscelyn@vdveen.ca>
Co-authored-by: Jean-Daniel Bancal <jdbancal@gmail.com>
Co-authored-by: Jean-Daniel Bancal <jean-daniel.bancal@ipht.fr>